# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MobiTrack - Phone & Accessory Business Management API for Zambia
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ── Sales ─────────────────────────────────────────────────
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List all sales
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Record a new sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete/void a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ── Expenses ──────────────────────────────────────────────
...
components:
  schemas:
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: integer
      required:
        - success
        - id

    # ── Product ────────────────────────────────────────────
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
    SupplierInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        contactName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string
      required:
        - name
    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
        contactName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string

    # ── Sale ──────────────────────────────────────────────
    Sale:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: number
        totalAmount:
          type: number
        totalCost:
          type: number
        profit:
          type: number
        paymentMethod:
          type: string
          description: "cash | mobile_money | bank_transfer | card"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - items
        - subtotal
        - discount
        - totalAmount
        - totalCost
        - profit
        - paymentMethod
        - createdAt
    SaleItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        unitCost:
          type: number
        lineTotal:
          type: number
        lineProfit:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - unitPrice
        - unitCost
        - lineTotal
        - lineProfit
    SaleInput:
      type: object
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
          minItems: 1
        discount:
          type: number
          minimum: 0
        paymentMethod:
          type: string
        notes:
          type: string
      required:
        - items
        - paymentMethod
    SaleItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          minimum: 0
      required:
        - productId
        - quantity
        - unitPrice

    # ── Expense ───────────────────────────────────────────
...