# lib/api-zod/src/generated/types/supplierUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MobiTrack - Phone & Accessory Business Management API for Zambia
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierUpdate {
  name?: string;
  contactName?: string;
  phone?: string;
  email?: string;
  address?: string;
  notes?: string;
}